This is the current final version I will create
Honestly this is basically unusable by anyone but me, it only works in windows, it has no graphical interface(only works on the command line), the code is very messy although understandable if you read very verryyyyy carefully. Basically I threw all the right coding practices out of the window(and I wasn't very good with those from the beginning) because I wanted to add the feature of allowing you to choose many chapters at once and just wanted to finish it quickly.
I will come back in a few months(maybe less than 1) and improve it for it to be at least usable for everybody in windows, and make it much more user friendly by adding a GUI. I would also like to improve the readability of the code by following the right practices in Python.